feat: update for WinUI packaged app
